# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Orbit Fundraising Intelligence API
  version: 1.0.0
  description: |
    Backend API for the Orbit autonomous fundraising platform.
    Serves the dashboard frontend and orchestrates CRM sync, AI agents, and webhooks.
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
  - url: https://api.orbitfundraising.com/api/v1
    description: Production
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /agents/{key}/queue:
    get:
      summary: Get agent donor queue (prioritized)
      tags: [Agents]
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string, enum: [VEO, VSO, VPGO, VCO] }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Prioritized donor list for agent
  /agents/{key}/activity:
    get:
      summary: Agent activity feed
      tags: [Agents]
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
        - name: since
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Activity log
  /agents/{key}/config:
    get:
      summary: Get agent configuration
      tags: [Agents]
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Agent config object

    put:
      summary: Save agent configuration
      tags: [Agents]
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: Full agent config (persona, cadence, thresholds, channels, etc.)
      responses:
        '200':
          description: Saved config
  /agents/{key}/run:
    post:
      summary: Trigger agent action on a donor (AI reasoning + outreach)
      tags: [Agents, AI]
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                donorId: { type: string }
                action:  { type: string, enum: [brief, compose, schedule, send] }
      responses:
        '200':
          description: Agent action result with AI reasoning trace

  # ── CAMPAIGNS ─────────────────────────────────────────────────────────────────
...